+++ packages/eslint-plugin-smarthr/rules/a11y-anchor-has-href-attribute/index.js
@@ -45,21 +45,11 @@ const OPTION = (() => {
 
 const ANCHOR_ELEMENT = 'JSXOpeningElement[name.name=/(Anchor|Link|^a)$/]'
 const HREF_ATTRIBUTE = `JSXAttribute[name.name=${OPTION.react_router ? '/^(href|to)$/' : '"href"'}]`
-const NULL_HREF_ATTRIBUTE_VALUES = `${HREF_ATTRIBUTE}:matches(${['#', ''].reduce((prev, v) => {
-  return `${prev},[value.type="Literal"][value.value="${v}"],[value.type="JSXExpressionContainer"][value.expression.value="${v}"]`
-}, '[value=null]')})`
-const NEXT_LINK_REGEX = /Link$/
-// HINT: next/link で `Link > a` という構造がありえるので直上のJSXElementを調べる
-const nextCheck = (node) => ((node.parent.parent.openingElement.name.name || '').test(NEXT_LINK_REGEX))
+const INVALID_HREF_SELECTOR = `${ANCHOR_ELEMENT}:has(${HREF_ATTRIBUTE}:matches([value=null],[value.type="Literal"][value.value="#"],[value.type="Literal"][value.value=""]))`
+const NO_HREF_SELECTOR = `${ANCHOR_ELEMENT}:not(:has(${HREF_ATTRIBUTE}))`
 
-const MESSAGE_SUFFIX = ` に href${OPTION.react_router ? '、もしくはto' : ''} 属性を正しく設定してください
- - 詳細: https://github.com/kufu/tamatebako/tree/master/packages/eslint-plugin-smarthr/rules/a11y-anchor-has-href-attribute
- - onClickなどでページ遷移する場合でもhref属性に遷移先のURIを設定してください
-   - Cmd + clickなどのキーボードショートカットに対応出来ます
- - onClickなどの動作がURLの変更を行わない場合、button要素でマークアップすることを検討してください
-   - href属性に空文字(""など)や '#' が設定されている場合、実質画面遷移を行わないため、同様にbutton要素でマークアップすることを検討してください
- - リンクが存在せず無効化されていることを表したい場合、href属性に undefined を設定してください
-   - button要素のdisabled属性が設定された場合に相当します`
+const NEXT_LINK_REGEX = /Link$/
+const nextCheck = (node) => NEXT_LINK_REGEX.test(node.parent.parent.openingElement.name.name || '')
 
 const SCHEMA = [
   {
@@ -82,20 +72,28 @@ module.exports = {
   create(context) {
     const option = context.options[0] || {}
     const spreadAttributeSelector = option.checkType === 'allow-spread-attributes' ? ':not(:has(JSXSpreadAttribute))' : ''
+
     const reporter = (node) => {
       context.report({
         node,
-        message: `${node.name.name}${MESSAGE_SUFFIX}`,
+        message: `${node.name.name} に href${OPTION.react_router ? '、もしくはto' : ''} 属性を正しく設定してください
+ - 詳細: https://github.com/kufu/tamatebako/tree/master/packages/eslint-plugin-smarthr/rules/a11y-anchor-has-href-attribute
+ - onClickなどでページ遷移する場合でもhref属性に遷移先のURIを設定してください
+   - Cmd + clickなどのキーボードショートカットに対応出来ます
+ - onClickなどの動作がURLの変更を行わない場合、button要素でマークアップすることを検討してください
+   - href属性に空文字(""など)や '#' が設定されている場合、実質画面遷移を行わないため、同様にbutton要素でマークアップすることを検討してください
+ - リンクが存在せず無効化されていることを表したい場合、href属性に undefined を設定してください
+   - button要素のdisabled属性が設定された場合に相当します`,
       })
     }
 
     return {
-      [`${ANCHOR_ELEMENT}:not(:has(${HREF_ATTRIBUTE}))${spreadAttributeSelector}`]: (node) => {
+      [`${NO_HREF_SELECTOR}${spreadAttributeSelector}`]: (node) => {
         if (!OPTION.nextjs || !nextCheck(node)) {
           reporter(node)
         }
       },
-      [`${ANCHOR_ELEMENT}:has(${NULL_HREF_ATTRIBUTE_VALUES})`]: reporter,
+      [INVALID_HREF_SELECTOR]: reporter,
     }
   },
 }