+++ packages/eslint-plugin-smarthr/rules/trim-props/index.js
@@ -1,5 +1,6 @@
 const SCHEMA = []
 
+const TRIM_REGEX = /^(['"`])\s*(.+?)\s*(['"`])$/
 
 /**
  * @type {import('@typescript-eslint/utils').TSESLint.RuleModule<''>}
@@ -11,20 +12,14 @@ module.exports = {
     fixable: 'whitespace',
   },
   create(context) {
-    const checker = (node) => {
-      return context.report({
-        node,
-        message: `属性に設定している文字列から先頭、末尾の空白文字を削除してください
- - 詳細: https://github.com/kufu/tamatebako/tree/master/packages/eslint-plugin-smarthr/rules/trim-props`,
-        fix: (fixer) => fixer.replaceText(node, context.sourceCode.getText(node).replace(/^('|"|`)\s+/, '$1').replace(/\s+('|"|`)$/, '$1')),
-      })
-    }
-
     return {
-      'JSXAttribute > Literal[value=/(^ | $)/]': checker,
-      'JSXAttribute > JSXExpressionContainer>Literal[value=/(^ | $)/]': checker,
-      'JSXAttribute > JSXExpressionContainer > TemplateLiteral > TemplateElement:matches(:first-child[value.raw=/^ /],:last-child[value.raw=/ $/])': (node) => {
-        checker(node.parent)
+      ':matches(JSXAttribute > Literal[value=/(^ | $)/], JSXAttribute > JSXExpressionContainer > TemplateLiteral:has(> TemplateElement:matches(:first-child[value.raw=/^ /],:last-child[value.raw=/ $/])))': (node) => {
+        context.report({
+          node,
+          message: `属性に設定している文字列から先頭、末尾の空白文字を削除してください
+ - 詳細: https://github.com/kufu/tamatebako/tree/master/packages/eslint-plugin-smarthr/rules/trim-props`,
+          fix: (fixer) => fixer.replaceText(node, context.sourceCode.getText(node).replace(TRIM_REGEX, '$1$2$3')),
+        })
       },
     }
   },