+++ packages/eslint-plugin-smarthr/test/a11y-aria-labelledby.js
@@ -23,7 +23,9 @@ ruleTester.run('a11y-aria-labelledby', rule, {
     { code: '<Any aria-labelledby={`${hoge} ${fuga}`} />' },
   ],
   invalid: [
+    // 文字列リテラル
     { code: `<Any aria-labelledby="hoge" />`, errors: [{ message: ERROR_MESSAGE }] },
+    // TemplateLiteral - 変数を含み、文字列リテラル部分がある
     { code: '<Any aria-labelledby={`hoge-${fuga}`} />', errors: [{ message: ERROR_MESSAGE }] },
     { code: '<Any aria-labelledby={`${"hoge"} ${fuga}`} />', errors: [{ message: ERROR_MESSAGE }] },
   ]